requirements.txt
openai>=2.0.0
python-dotenv>=1.1.1
requests>=2.32.5
